chore: clippy --fix across workspace + tarball_str &Path
Mechanical clippy fixes (no behavior change): collapse if-else
(prompt_ratio), char pattern match, useless .into()/cast, contains vs
iter().any(), is_some_and, trim before split_whitespace. Plus
tarball_str &PathBuf -> &Path (ptr_arg). Fixes ~20 pre-existing
warnings; one bin ptr_arg warning remains out of clippy --fix reach.

+++ crates/neomind-core/src/eventbus.rs
@@ -165,7 +165,7 @@ impl EventBusReceiver {
             match self.rx.recv().await {
                 Ok(event) => return Some(event),
                 Err(broadcast::error::RecvError::Lagged(n)) => {
-                    let total = self.dropped.fetch_add(n as u64, Ordering::Relaxed) + n as u64;
+                    let total = self.dropped.fetch_add(n, Ordering::Relaxed) + n;
                     tracing::warn!(
                         dropped_total = total,
                         this_batch = n,
@@ -240,7 +240,7 @@ where
                     // Event didn't match filter, continue waiting
                 }
                 Err(broadcast::error::RecvError::Lagged(n)) => {
-                    let total = self.dropped.fetch_add(n as u64, Ordering::Relaxed) + n as u64;
+                    let total = self.dropped.fetch_add(n, Ordering::Relaxed) + n;
                     tracing::warn!(
                         dropped_total = total,
                         this_batch = n,
